requirements.txt
requests~=2.31.0
urllib3~=2.0.7
Pillow~=10.1.0
PyGetWindow~=0.0.9